+++ docs/release_notes.rst
@@ -12,8 +12,8 @@ Features in the 0.6.x series of releases are focused on expanding functionality 
 0.6.1
 `````
 
-    * Add iff and implies Boolean operators
-    * Add :data:`is_cnf <tt.expressions.bexpr.BooleanExpression.is_cnf>` attribute to :class:`BooleanExpression <tt.expressions.bexpr.BooleanExpression>`
+    * Add iff (iff, ->) and implies (impl, <->) Boolean operators
+    * Add :data:`is_cnf <tt.expressions.bexpr.BooleanExpression.is_cnf>` and :data:`is_dnf <tt.expressions.bexpr.BooleanExpression.is_dnf>` attributes to :class:`BooleanExpression <tt.expressions.bexpr.BooleanExpression>`
 
 0.6.0
 `````